amplify/backend/function/dungeongen7bea1f1c/src/pyproject.toml
[project]
name = "dungeongen-backend"
version = "0.1.0"
description = "DungeonGen Backend API"
authors = [
    {name = "DungeonGen Team"}
]
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-restx>=1.3.0",
    "python-dotenv>=1.1.1",
    "groq>=0.4.0",
    "langchain>=0.1.0",
    "langchain-groq>=0.0.1",
    "pydantic>=2.0.0",
    "deprecated>=1.2.13",
    "json-repair>=0.25.0",
    "numpy>=1.24.0",
    "networkx>=3.5",
    "pyjwt>=2.8.0",
    "bcrypt>=4.0.1",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.12.11",
    "black>=25.1.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
target-version = ['py311']
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
